+++ lib/ClangImporter/CMakeLists.txt
@@ -10,6 +10,7 @@ handle_gyb_sources(
 
 add_swift_library(swiftClangImporter STATIC
   CFTypeInfo.cpp
+  ClangAdapter.cpp
   ClangDiagnosticConsumer.cpp
   ClangImporter.cpp
   IAMInference.cpp
