+++ lib/ClangImporter/ImportType.cpp
@@ -597,7 +597,7 @@ namespace {
       Type mappedType = getAdjustedTypeDeclReferenceType(decl);
       ImportHint hint = ImportHint::None;
 
-      if (Impl.getSwiftNewtypeAttr(type->getDecl(), /*useSwift2Name=*/false)) {
+      if (getSwiftNewtypeAttr(type->getDecl(), /*useSwift2Name=*/false)) {
         if (isCFTypeDecl(type->getDecl()))
           hint = ImportHint::SwiftNewtypeFromCFPointer;
         else
@@ -1379,18 +1379,6 @@ Type ClangImporter::Implementation::importType(clang::QualType type,
                                      optionality);
 }
 
-bool ClangImporter::Implementation::isNSString(const clang::Type *type) {
-  if (auto ptrType = type->getAs<clang::ObjCObjectPointerType>())
-    if (auto interfaceType = ptrType->getInterfaceType())
-      if (interfaceType->getDecl()->getName() == "NSString")
-        return true;
-  return false;
-}
-
-bool ClangImporter::Implementation::isNSString(clang::QualType qt) {
-  return qt.getTypePtrOrNull() && isNSString(qt.getTypePtrOrNull());
-}
-
 bool ClangImporter::Implementation::shouldImportGlobalAsLet(
        clang::QualType type)
 {
@@ -1461,39 +1449,6 @@ Type ClangImporter::Implementation::importPropertyType(
                     /*isFullyBridgeable*/ true, optionality);
 }
 
-/// Get a bit vector indicating which arguments are non-null for a
-/// given function or method.
-llvm::SmallBitVector ClangImporter::Implementation::getNonNullArgs(
-                       const clang::Decl *decl,
-                       ArrayRef<const clang::ParmVarDecl *> params) {
-  llvm::SmallBitVector result;
-  if (!decl)
-    return result;
-
-  for (const auto *nonnull : decl->specific_attrs<clang::NonNullAttr>()) {
-    if (!nonnull->args_size()) {
-      // Easy case: all pointer arguments are non-null.
-      if (result.empty())
-        result.resize(params.size(), true);
-      else
-        result.set(0, params.size());
-
-      return result;
-    }
-
-    // Mark each of the listed parameters as non-null.
-    if (result.empty())
-      result.resize(params.size(), false);
-
-    for (unsigned idx : nonnull->args()) {
-      if (idx < result.size())
-        result.set(idx);
-    }
-  }
-
-  return result;
-}
-
 /// Apply the @noescape attribute
 static Type applyNoEscape(Type type) {
   // Recurse into optional types.
@@ -1524,7 +1479,7 @@ static OptionalTypeKind getParamOptionality(
 
   // If nullability is available on the type, use it.
   if (auto nullability = param->getType()->getNullability(clangCtx)) {
-    return ClangImporter::Implementation::translateNullability(*nullability);
+    return translateNullability(*nullability);
   }
 
   // If it's known non-null, use that.
@@ -1693,302 +1648,6 @@ static bool isObjCMethodResultAudited(const clang::Decl *decl) {
           decl->hasAttr<clang::ObjCReturnsInnerPointerAttr>());
 }
 
-/// Determine whether this is the name of a collection with a single
-/// element type.
-static bool isCollectionName(StringRef typeName) {
-  auto lastWord = camel_case::getLastWord(typeName);
-  return lastWord == "Array" || lastWord == "Set";
-}
-
-/// Retrieve the name of the given Clang type for use when omitting
-/// needless words.
-OmissionTypeName ClangImporter::Implementation::getClangTypeNameForOmission(
-                   clang::ASTContext &ctx, clang::QualType type) {
-  if (type.isNull())
-    return OmissionTypeName();
-
-  // Dig through the type, looking for a typedef-name and stripping
-  // references along the way.
-  StringRef lastTypedefName;
-  do {
-    // The name of a typedef-name.
-    auto typePtr = type.getTypePtr();
-    if (auto typedefType = dyn_cast<clang::TypedefType>(typePtr)) {
-      auto name = typedefType->getDecl()->getName();
-
-      // Objective-C selector type.
-      if (ctx.hasSameUnqualifiedType(type, ctx.getObjCSelType()) &&
-          name == "SEL")
-        return "Selector";
-
-      // Objective-C "id" type.
-      if (type->isObjCIdType() && name == "id")
-        return "Object";
-
-      // Objective-C "Class" type.
-      if (type->isObjCClassType() && name == "Class")
-        return "Class";
-
-      // Objective-C "BOOL" type.
-      if (name == "BOOL")
-        return OmissionTypeName("Bool", OmissionTypeFlags::Boolean);
-
-      // If this is an imported CF type, use that name.
-      StringRef CFName = getCFTypeName(typedefType->getDecl());
-      if (!CFName.empty())
-        return CFName;
-
-      // If we have NS(U)Integer or CGFloat, return it.
-      if (name == "NSInteger" || name == "NSUInteger" || name == "CGFloat")
-        return name;
-
-      // If it's a collection name and of pointer type, call it an
-      // array of the pointee type.
-      if (isCollectionName(name)) {
-        if (auto ptrType = type->getAs<clang::PointerType>()) {
-          return OmissionTypeName(
-                   name, None, 
-                 getClangTypeNameForOmission(ctx, ptrType->getPointeeType())
-                   .Name);
-        }
-      }
-
-      // Otherwise, desugar one level...
-      lastTypedefName = name;
-      type = typedefType->getDecl()->getUnderlyingType();
-      continue;
-    }
-
-    // For array types, convert the element type and treat this an as array.
-    if (auto arrayType = dyn_cast<clang::ArrayType>(typePtr)) {
-      return OmissionTypeName(
-               "Array", None, 
-               getClangTypeNameForOmission(ctx, arrayType->getElementType())
-                 .Name);
-    }
-
-    // Look through reference types.
-    if (auto refType = dyn_cast<clang::ReferenceType>(typePtr)) {
-      type = refType->getPointeeTypeAsWritten();
-      continue;
-    }
-
-    // Look through pointer types.
-    if (auto ptrType = dyn_cast<clang::PointerType>(typePtr)) {
-      type = ptrType->getPointeeType();
-      continue;
-    }
-
-    // Try to desugar one level...
-    clang::QualType desugared = type.getSingleStepDesugaredType(ctx);
-    if (desugared.getTypePtr() == type.getTypePtr())
-      break;
-
-    type = desugared;
-  } while (true);
-
-  // Objective-C object pointers.
-  if (auto objcObjectPtr = type->getAs<clang::ObjCObjectPointerType>()) {
-    auto objcClass = objcObjectPtr->getInterfaceDecl();
-
-    // For id<Proto> or NSObject<Proto>, retrieve the name of "Proto".
-    if (objcObjectPtr->getNumProtocols() == 1 &&
-        (!objcClass || objcClass->getName() == "NSObject"))
-      return (*objcObjectPtr->qual_begin())->getName();
-
-    // If there is a class, use it.
-    if (objcClass) {
-      // If this isn't the name of an Objective-C collection, we're done.
-      auto className = objcClass->getName();
-      if (!isCollectionName(className))
-        return className;
-
-      // If we don't have type parameters, use the prefix of the type
-      // name as the collection element type.
-      if (objcClass && !objcClass->getTypeParamList()) {
-        unsigned lastWordSize = camel_case::getLastWord(className).size();
-        StringRef elementName =
-          className.substr(0, className.size() - lastWordSize);
-        return OmissionTypeName(className, None, elementName);
-      }
-
-      // If we don't have type arguments, the collection element type
-      // is "Object".
-      auto typeArgs = objcObjectPtr->getTypeArgs();
-      if (typeArgs.empty())
-        return OmissionTypeName(className, None, "Object");
-
-      return OmissionTypeName(className, None,
-                              getClangTypeNameForOmission(ctx,
-                                                          typeArgs[0]).Name);
-    }
-
-    // Objective-C "id" type.
-    if (objcObjectPtr->isObjCIdType())
-      return "Object";
-
-    // Objective-C "Class" type.
-    if (objcObjectPtr->isObjCClassType())
-      return "Class";
-
-    return StringRef();
-  }
-
-  // Handle builtin types by importing them and getting the Swift name.
-  if (auto builtinTy = type->getAs<clang::BuiltinType>()) {
-    // Names of integer types.
-    static const char *intTypeNames[] = {
-      "UInt8",
-      "UInt16",
-      "UInt32",
-      "UInt64",
-      "UInt128"
-    };
-
-    /// Retrieve the name for an integer type based on its size.
-    auto getIntTypeName = [&](bool isSigned) -> StringRef {
-      switch (ctx.getTypeSize(builtinTy)) {
-      case 8: return StringRef(intTypeNames[0]).substr(isSigned ? 1 : 0);
-      case 16: return StringRef(intTypeNames[1]).substr(isSigned ? 1 : 0);
-      case 32: return StringRef(intTypeNames[2]).substr(isSigned ? 1 : 0);
-      case 64: return StringRef(intTypeNames[3]).substr(isSigned ? 1 : 0);
-      case 128: return StringRef(intTypeNames[4]).substr(isSigned ? 1 : 0);
-      default: llvm_unreachable("bad integer type size");
-      }
-    };
-
-    switch (builtinTy->getKind()) {
-    case clang::BuiltinType::Void:
-      return "Void";
-
-    case clang::BuiltinType::Bool:
-      return OmissionTypeName("Bool", OmissionTypeFlags::Boolean);
-
-    case clang::BuiltinType::Float:
-      return "Float";
-
-    case clang::BuiltinType::Double:
-      return "Double";
-
-    case clang::BuiltinType::Char16:
-      return "UInt16";
-
-    case clang::BuiltinType::Char32:
-      return "UnicodeScalar";
-
-    case clang::BuiltinType::Char_U:
-    case clang::BuiltinType::UChar:
-    case clang::BuiltinType::UShort:
-    case clang::BuiltinType::UInt:
-    case clang::BuiltinType::ULong:
-    case clang::BuiltinType::ULongLong:
-    case clang::BuiltinType::UInt128:
-    case clang::BuiltinType::WChar_U:
-      return getIntTypeName(false);
-
-    case clang::BuiltinType::Char_S:
-    case clang::BuiltinType::SChar:
-    case clang::BuiltinType::Short:
-    case clang::BuiltinType::Int:
-    case clang::BuiltinType::Long:
-    case clang::BuiltinType::LongLong:
-    case clang::BuiltinType::Int128:
-    case clang::BuiltinType::WChar_S:
-      return getIntTypeName(true);
-
-    // Types that cannot be mapped into Swift, and probably won't ever be.
-    case clang::BuiltinType::Dependent:
-    case clang::BuiltinType::ARCUnbridgedCast:
-    case clang::BuiltinType::BoundMember:
-    case clang::BuiltinType::BuiltinFn:
-    case clang::BuiltinType::Overload:
-    case clang::BuiltinType::PseudoObject:
-    case clang::BuiltinType::UnknownAny:
-      return OmissionTypeName();
-
-    // FIXME: Types that can be mapped, but aren't yet.
-    case clang::BuiltinType::Half:
-    case clang::BuiltinType::LongDouble:
-    case clang::BuiltinType::NullPtr:
-      return OmissionTypeName();
-
-    // Objective-C types that aren't mapped directly; rather, pointers to
-    // these types will be mapped.
-    case clang::BuiltinType::ObjCClass:
-    case clang::BuiltinType::ObjCId:
-    case clang::BuiltinType::ObjCSel:
-      return OmissionTypeName();
-
-    // OpenCL types that don't have Swift equivalents.
-    case clang::BuiltinType::OCLImage1dRO:
-    case clang::BuiltinType::OCLImage1dRW:
-    case clang::BuiltinType::OCLImage1dWO:
-    case clang::BuiltinType::OCLImage1dArrayRO:
-    case clang::BuiltinType::OCLImage1dArrayRW:
-    case clang::BuiltinType::OCLImage1dArrayWO:
-    case clang::BuiltinType::OCLImage1dBufferRO:
-    case clang::BuiltinType::OCLImage1dBufferRW:
-    case clang::BuiltinType::OCLImage1dBufferWO:
-    case clang::BuiltinType::OCLImage2dRO:
-    case clang::BuiltinType::OCLImage2dRW:
-    case clang::BuiltinType::OCLImage2dWO:
-    case clang::BuiltinType::OCLImage2dArrayRO:
-    case clang::BuiltinType::OCLImage2dArrayRW:
-    case clang::BuiltinType::OCLImage2dArrayWO:
-    case clang::BuiltinType::OCLImage2dDepthRO:
-    case clang::BuiltinType::OCLImage2dDepthRW:
-    case clang::BuiltinType::OCLImage2dDepthWO:
-    case clang::BuiltinType::OCLImage2dArrayDepthRO:
-    case clang::BuiltinType::OCLImage2dArrayDepthRW:
-    case clang::BuiltinType::OCLImage2dArrayDepthWO:
-    case clang::BuiltinType::OCLImage2dMSAARO:
-    case clang::BuiltinType::OCLImage2dMSAARW:
-    case clang::BuiltinType::OCLImage2dMSAAWO:
-    case clang::BuiltinType::OCLImage2dArrayMSAARO:
-    case clang::BuiltinType::OCLImage2dArrayMSAARW:
-    case clang::BuiltinType::OCLImage2dArrayMSAAWO:
-    case clang::BuiltinType::OCLImage2dMSAADepthRO:
-    case clang::BuiltinType::OCLImage2dMSAADepthRW:
-    case clang::BuiltinType::OCLImage2dMSAADepthWO:
-    case clang::BuiltinType::OCLImage2dArrayMSAADepthRO:
-    case clang::BuiltinType::OCLImage2dArrayMSAADepthRW:
-    case clang::BuiltinType::OCLImage2dArrayMSAADepthWO:
-    case clang::BuiltinType::OCLImage3dRO:
-    case clang::BuiltinType::OCLImage3dRW:
-    case clang::BuiltinType::OCLImage3dWO:
-    case clang::BuiltinType::OCLSampler:
-    case clang::BuiltinType::OCLEvent:
-    case clang::BuiltinType::OCLClkEvent:
-    case clang::BuiltinType::OCLQueue:
-    case clang::BuiltinType::OCLNDRange:
-    case clang::BuiltinType::OCLReserveID:
-      return OmissionTypeName();
-
-    // OpenMP types that don't have Swift equivalents.
-    case clang::BuiltinType::OMPArraySection:
-      return OmissionTypeName();
-    }
-  }
-
-  // Tag types.
-  if (auto tagType = type->getAs<clang::TagType>()) {
-    if (tagType->getDecl()->getName().empty())
-      return lastTypedefName;
-
-    return tagType->getDecl()->getName();
-  }
-
-  // Block pointers.
-  if (type->getAs<clang::BlockPointerType>())
-    return OmissionTypeName("Block", OmissionTypeFlags::Function);
-
-  // Function pointers.
-  if (type->isFunctionType())
-    return OmissionTypeName("Function", OmissionTypeFlags::Function);
-
-  return StringRef();
-}
-
 /// Attempt to omit needless words from the given function name.
 bool ClangImporter::Implementation::omitNeedlessWordsInFunctionName(
        clang::Sema &clangSema,
@@ -2058,35 +1717,6 @@ bool ClangImporter::Implementation::omitNeedlessWordsInFunctionName(
                            getClangTypeNameForOmission(clangCtx, contextType),
                            paramTypes, returnsSelf, /*isProperty=*/false,
                            allPropertyNames, scratch);
-}
-
-/// Retrieve the instance type of the given Clang declaration context.
-clang::QualType ClangImporter::Implementation::getClangDeclContextType(
-                  const clang::DeclContext *dc) {
-  auto &ctx = dc->getParentASTContext();
-  if (auto objcClass = dyn_cast<clang::ObjCInterfaceDecl>(dc))
-    return ctx.getObjCObjectPointerType(ctx.getObjCInterfaceType(objcClass));
-
-  if (auto objcCategory = dyn_cast<clang::ObjCCategoryDecl>(dc)) {
-    if (objcCategory->isInvalidDecl()) return clang::QualType();
-    
-    return ctx.getObjCObjectPointerType(
-             ctx.getObjCInterfaceType(
-               objcCategory->getClassInterface()));
-  }
-
-  if (auto constProto = dyn_cast<clang::ObjCProtocolDecl>(dc)) {
-    auto proto = const_cast<clang::ObjCProtocolDecl *>(constProto);
-    auto type = ctx.getObjCObjectType(ctx.ObjCBuiltinIdTy, { }, { proto },
-                                      false);
-    return ctx.getObjCObjectPointerType(type);
-  }
-
-  if (auto tag = dyn_cast<clang::TagDecl>(dc)) {
-    return ctx.getTagDeclType(tag);
-  }
-
-  return clang::QualType();
 }
 
 DefaultArgumentKind ClangImporter::Implementation::inferDefaultArgument(
